// src/app/views/prj-planning/model/task-attachment.ts
export class TaskAttachment  {
  attachmentType?: string;
  oid?: string;
  fileName?: string;
  fileOid?: string;
  filePath?: string;
  fileType?: string;
  tag?: string;
  description?: string;
  createdBy?:string;
  fileDownloadUri?: string;
  requestOid?: string;
  size?: number;
  title?: string;
}